FastAPI/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
Pillow==10.0.0

